+++ SonarApp/app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.example.sonarapp"
+        applicationId "com.golovkobalak.sonarapp"
         minSdkVersion 23
         targetSdkVersion 29
         versionCode 1
