+++ SonarApp/app/src/androidTest/java/com/golovkobalak/sonarapp/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.sonarapp;
+package com.golovkobalak.sonarapp;
 
 import android.content.Context;
