+++ SonarApp/app/src/main/java/com/golovkobalak/sonarapp/SonarActivity.java
@@ -1,4 +1,4 @@
-package com.example.sonarapp;
+package com.golovkobalak.sonarapp;
 
 import android.Manifest;
 import android.app.Activity;
