+++ SonarApp/app/src/main/java/com/golovkobalak/sonarapp/ui/dashboard/DashboardViewModel.java
@@ -1,16 +1,16 @@
-package com.example.sonarapp.ui.home;
+package com.golovkobalak.sonarapp.ui.dashboard;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class DashboardViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public DashboardViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is dashboard fragment");
     }
 
     public LiveData<String> getText() {
