+++ SonarApp/app/src/main/java/com/golovkobalak/sonarapp/ui/home/HomeViewModel.java
@@ -1,16 +1,16 @@
-package com.example.sonarapp.ui.dashboard;
+package com.golovkobalak.sonarapp.ui.home;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
